requirements.txt
langchain_text_splitters
langchain-google-genai
langchain_community
langchain_chroma
langchain_ollama
langchain
langgraph-checkpoint-sqlite
langgraph
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
instructor[google-genai]
instructor
python-dotenv
line-bot-sdk
fastapi
uvicorn
pandas